+++ now_lms/vistas/groups.py
@@ -3,13 +3,12 @@
 # ---------------------------------------------------------------------------------------
 # Standard library
 # ---------------------------------------------------------------------------------------
-from datetime import datetime
 
 # ---------------------------------------------------------------------------------------
 # Third-party libraries
 # ---------------------------------------------------------------------------------------
-from flask import Blueprint, flash, redirect, render_template, request, url_for
-from flask_login import current_user, login_required
+from flask import Blueprint, flash, redirect, render_template, request
+from flask_login import login_required
 from sqlalchemy.exc import OperationalError
 
 # ---------------------------------------------------------------------------------------
@@ -17,7 +16,7 @@ from sqlalchemy.exc import OperationalError
 # ---------------------------------------------------------------------------------------
 from now_lms.auth import perfil_requerido
 from now_lms.config import DIRECTORIO_PLANTILLAS
-from now_lms.db import UsuarioGrupo, UsuarioGrupoTutor, database
+from now_lms.db import UsuarioGrupo, database
 from now_lms.forms import GrupoForm
 
 group = Blueprint("group", __name__, template_folder=DIRECTORIO_PLANTILLAS)
@@ -47,29 +46,3 @@ def nuevo_grupo():
             return redirect("/new_group")
     else:
         return render_template("admin/grupos/nuevo.html", form=form)
-
-
-@group.route(
-    "/group/set_tutor",
-    methods=[
-        "POST",
-    ],
-)
-@login_required
-@perfil_requerido("admin")
-def agrega_tutor_a_grupo():
-    """Asigna como tutor de grupo a un usuario."""
-    id_ = request.args.get("id", type=str)
-    registro = UsuarioGrupoTutor(
-        grupo=id_, usuario=request.form["usuario"], creado_por=current_user.usuario, creado=datetime.now()
-    )
-    database.session.add(registro)
-    url_grupo = url_for("grupo", id=id_)
-    try:
-        database.session.commit()
-        flash("Usuario Agregado Correctamente.", "success")
-        return redirect(url_grupo)
-    except OperationalError:  # pragma: no cover
-        database.session.rollback()
-        flash("No se pudo agregar al usuario.", "warning")
-        return redirect(url_grupo)