+++ CHANGELOG.md
@@ -6,6 +6,26 @@ All notable changes to this project will be documented in this file.
 - This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - This project also follows [Conventional Commits](https://www.conventionalcommits.org/), using prefixes such as:
 
+## [release: v0.0.1b12.dev20250823] - 2025-08-23
+
+### Notes:
+ - Updated docs
+
+## [release: v0.0.1b11.dev20250822] - 2025-08-22
+
+### Fixed:
+ - Fixed a bug in the certificate implementation
+ - Fixed a bug resources creation
+
+### Notes
+
+ - Included pylint as defensive tool
+ - 698 tests passed / 83% code coverage
+
+## [v0.0.1b10.dev20250822] -2025-08-22
+
+### Unreleased
+
 ## [v0.0.1b9.dev20250818] -2025-08-18
 
 ### Added